Refactored to move the code that uses Markup interface together with the definition.
