+++ markup.go
@@ -1,13 +1,179 @@
 package main
 
+import (
+	"bytes"
+	"fmt"
+	"github.com/yvasiyarov/swagger/parser"
+	"log"
+	"os"
+	"path"
+	"sort"
+	"strings"
+)
+
 type Markup interface {
-	anchor(anchorName string) string
 	sectionHeader(level int, text string) string
 	bulletedItem(level int, text string) string
 	numberedItem(level int, text string) string
+	anchor(anchorName string) string
 	link(anchorName, linkText string) string
 	tableHeader(tableTitle string) string
-	tableFooter() string
 	tableHeaderRow(args ...string) string
 	tableRow(args ...string) string
+	tableFooter() string
+}
+
+func generateMarkup(parser *parser.Parser, markup Markup, fileExtension string) {
+	var filename string
+	if *outputSpec == "" {
+		filename = path.Join("./", "API", fileExtension)
+	} else {
+		filename = path.Join(*outputSpec)
+	}
+	fd, err := os.Create(filename)
+	if err != nil {
+		log.Fatalf("Can not create document file: %v\n", err)
+	}
+	defer fd.Close()
+
+	var buf bytes.Buffer
+
+	/***************************************************************
+	* Overall API
+	***************************************************************/
+	buf.WriteString(markup.sectionHeader(1, parser.Listing.Infos.Title))
+	buf.WriteString(fmt.Sprintf("%s\n\n", parser.Listing.Infos.Description))
+
+	/***************************************************************
+	* Table of Contents (List of Sub-APIs)
+	***************************************************************/
+	buf.WriteString("Table of Contents\n\n")
+	subApiKeys, subApiKeyIndex := alphabeticalKeysOfSubApis(parser.Listing.Apis)
+	for _, subApiKey := range subApiKeys {
+		buf.WriteString(markup.numberedItem(1, markup.link(subApiKey, parser.Listing.Apis[subApiKeyIndex[subApiKey]].Description)))
+	}
+	buf.WriteString("\n")
+
+	for _, apiKey := range alphabeticalKeysOfApiDeclaration(parser.TopLevelApis) {
+
+		apiDescription := parser.TopLevelApis[apiKey]
+		/***************************************************************
+		* Sub-API Specifications
+		***************************************************************/
+		buf.WriteString(markup.anchor(apiKey))
+		buf.WriteString(markup.sectionHeader(2, apiKey))
+
+		buf.WriteString(markup.tableHeader(""))
+		buf.WriteString(markup.tableHeaderRow("Specification", "Value"))
+		buf.WriteString(markup.tableRow("Resource Path", apiDescription.ResourcePath))
+		buf.WriteString(markup.tableRow("API Version", apiDescription.ApiVersion))
+		buf.WriteString(markup.tableRow("BasePath for the API", apiDescription.BasePath))
+		buf.WriteString(markup.tableRow("Consumes", strings.Join(apiDescription.Consumes, ", ")))
+		buf.WriteString(markup.tableRow("Produces", strings.Join(apiDescription.Produces, ", ")))
+		buf.WriteString(markup.tableFooter())
+
+		/***************************************************************
+		* Sub-API Operations (Summary)
+		***************************************************************/
+		buf.WriteString("\n")
+		buf.WriteString(markup.sectionHeader(3, "Operations"))
+		buf.WriteString("\n")
+
+		buf.WriteString(markup.tableHeader(""))
+		buf.WriteString(markup.tableHeaderRow("Resource Path", "Operation", "Description"))
+		for _, subapi := range apiDescription.Apis {
+			for _, op := range subapi.Operations {
+				pathString := strings.Replace(strings.Replace(subapi.Path, "{", "\\{", -1), "}", "\\}", -1)
+				buf.WriteString(markup.tableRow(pathString, markup.link(op.Nickname, op.HttpMethod), op.Summary))
+			}
+		}
+		buf.WriteString(markup.tableFooter())
+		buf.WriteString("\n")
+
+		/***************************************************************
+		* Sub-API Operations (Details)
+		***************************************************************/
+		for _, subapi := range apiDescription.Apis {
+			for _, op := range subapi.Operations {
+				buf.WriteString("\n")
+				operationString := fmt.Sprintf("%s (%s)", strings.Replace(strings.Replace(subapi.Path, "{", "\\{", -1), "}", "\\}", -1), op.HttpMethod)
+				buf.WriteString(markup.anchor(op.Nickname))
+				buf.WriteString(markup.sectionHeader(4, "API: "+operationString))
+				buf.WriteString("\n\n" + op.Summary + "\n\n\n")
+
+				if len(op.Parameters) > 0 {
+					buf.WriteString(markup.tableHeader(""))
+					buf.WriteString(markup.tableHeaderRow("Param Name", "Param Type", "Data Type", "Description", "Required?"))
+					for _, param := range op.Parameters {
+						isRequired := ""
+						if param.Required {
+							isRequired = "Yes"
+						}
+						buf.WriteString(markup.tableRow(param.Name, param.ParamType, param.DataType, param.Description, isRequired))
+					}
+					buf.WriteString(markup.tableFooter())
+				}
+
+				if len(op.ResponseMessages) > 0 {
+					buf.WriteString(markup.tableHeader(""))
+					buf.WriteString(markup.tableHeaderRow("Code", "Message", "Model"))
+					for _, msg := range op.ResponseMessages {
+						shortName := shortModelName(msg.ResponseModel)
+						modelText := shortName
+						if msg.ResponseModel != shortName {
+							modelText = markup.link(msg.ResponseModel, shortName)
+						}
+						buf.WriteString(markup.tableRow(fmt.Sprintf("%v", msg.Code), msg.Message, modelText))
+					}
+					buf.WriteString(markup.tableFooter())
+				}
+			}
+		}
+		buf.WriteString("\n")
+
+		/***************************************************************
+		* Models
+		***************************************************************/
+		buf.WriteString("\n")
+		buf.WriteString(markup.sectionHeader(3, "Models"))
+		buf.WriteString("\n")
+
+		for modelKey, model := range apiDescription.Models {
+			buf.WriteString(markup.anchor(modelKey))
+			buf.WriteString(markup.sectionHeader(4, shortModelName(modelKey)))
+			buf.WriteString(markup.tableHeader(""))
+			buf.WriteString(markup.tableHeaderRow("Field Name", "Field Type", "Description"))
+			for fieldName, fieldProps := range model.Properties {
+				buf.WriteString(markup.tableRow(fieldName, fieldProps.Type, fieldProps.Description))
+			}
+			buf.WriteString(markup.tableFooter())
+			buf.WriteString("\nNote: These fields are listed in random order (for now).\n")
+		}
+		buf.WriteString("\n")
+
+	}
+
+	fd.WriteString(buf.String())
+}
+
+func alphabeticalKeysOfSubApis(refs []*parser.ApiRef) ([]string, map[string]int) {
+	index := map[string]int{}
+	keys := make([]string, len(refs))
+	for i, ref := range refs {
+		subApiKey := ref.Path[1:]
+		keys[i] = subApiKey
+		index[subApiKey] = i
+	}
+	sort.Strings(keys)
+	return keys, index
+}
+func alphabeticalKeysOfApiDeclaration(m map[string]*parser.ApiDeclaration) []string {
+	keys := make([]string, len(m))
+	i := 0
+	for key, _ := range m {
+		keys[i] = key
+		i++
+	}
+	sort.Strings(keys)
+	return keys
 }
